- api key parameter has more precedence than environment variable - moved build tests to own file
