+++ tests/utils.py
@@ -1,0 +1,14 @@
+import os
+
+DEFAULT_MATCHES_SIZE = 100
+LANGUAGE_PAR = 'language=en_us'
+STEAM_ID_PAR = 'key=' + os.environ.get('D2_API_KEY')
+
+
+def convert_to_64_bit(number):
+    # Yes we should put this in the API - will be used to parse steam names
+    return number + 76561197960265728
+
+
+def request_pars(*args):
+    return '?' + '&'.join(args)
